Modification des types quantitées: int -> float
Modification du type d'élement en Sortie:
PF -> Element

Ajout des méthodes et classes pour l'ajout de chaine de production par
lecture de fichier csv:
-FichierCSVChaineDeProduction

+++ src/core/Stockage.java
@@ -6,14 +6,14 @@ import java.util.Iterator;
 public class Stockage {
 	private String codeStockage;
 	private String nom;
-	private HashMap<Element, Integer> stock;
+	private HashMap<Element, Float> stock;
 	public Stockage(String codeStockage, String nom) {
 		this.codeStockage = codeStockage;
 		this.nom=nom;
-		this.stock = new HashMap<Element, Integer>();
+		this.stock = new HashMap<Element, Float>();
 	}
 	
-	public void ajouterElementPourStockage(Element elem, int quantitee) {
+	public void ajouterElementPourStockage(Element elem, float quantitee) {
 		if(this.stock.containsKey(elem)) {
 			this.stock.put(elem, this.stock.get(elem)+quantitee);
 		}else {
@@ -24,7 +24,7 @@ public class Stockage {
 		this.stock.remove(elem);
 	}
 	
-	public void retirerElementPourStockage(Element elem, int quantitee) {
+	public void retirerElementPourStockage(Element elem, float quantitee) {
 		if(this.stock.containsKey(elem)) {
 			if(this.stock.get(elem)-quantitee >= 0) {
 				this.stock.put(elem, this.stock.get(elem)-quantitee);
@@ -38,7 +38,7 @@ public class Stockage {
 		String src = "stockage : :" +this.codeStockage + " - " + this.nom + ":\n" ;
 		Iterator it = this.stock.keySet().iterator();
 		Element elem;
-		Integer quantitee;
+		float quantitee;
 		while(it.hasNext()) {
 			elem = (Element) it.next(); 
 			quantitee = this.stock.get(elem);
